scan: new module for Scan trait and SyntaxError
We want to generalize the Scan trait so that it no longer depends on the
zone file parser (master::scan::Scanner), to facilitate record
deserialization outside of zone files (e.g. for parsing command-line
arguments).
