+++ src/base/rdata.rs
@@ -29,9 +29,9 @@ use super::octets::{
     ShortBuf,
 };
 #[cfg(feature = "master")]
-use crate::master::scan::{
-    CharSource, Scan, ScanError, Scanner, SyntaxError,
-};
+use crate::master::scan::{CharSource, Scanner};
+#[cfg(feature = "scan")]
+use crate::scan::{Scan, ScanError, SyntaxError};
 #[cfg(feature = "master")]
 use bytes::{BufMut, Bytes, BytesMut};
 use core::cmp::Ordering;
