+++ src/lib.rs
@@ -89,6 +89,7 @@ pub mod base;
 pub mod master;
 pub mod rdata;
 pub mod resolv;
+pub mod scan;
 pub mod sign;
 pub mod test;
 pub mod tsig;
