+++ src/master/reader.rs
@@ -1,8 +1,9 @@
 use super::entry::{Entry, MasterRecord};
-use super::scan::{CharSource, Pos, ScanError, Scanner};
+use super::scan::{CharSource, Pos, Scanner};
 use super::source::Utf8File;
 use crate::base::iana::Class;
 use crate::base::name::Dname;
+use crate::scan::ScanError;
 use bytes::Bytes;
 use std::fmt;
 use std::io;
@@ -130,7 +131,7 @@ impl fmt::Display for ReaderItem {
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::master::scan::ScanError;
+    use crate::scan::ScanError;
 
     #[test]
     fn print() {
