+++ src/rdata/rfc6672.rs
@@ -7,7 +7,9 @@ use crate::base::octets::{
 };
 use crate::base::rdata::RtypeRecordData;
 #[cfg(feature = "master")]
-use crate::master::scan::{CharSource, Scan, ScanError, Scanner};
+use crate::master::scan::{CharSource, Scanner};
+#[cfg(feature = "scan")]
+use crate::scan::{Scan, ScanError};
 use core::cmp::Ordering;
 use core::str::FromStr;
 use core::{fmt, hash, ops};
