+++ src/rdata/rfc7344.rs
@@ -9,7 +9,9 @@ use crate::base::octets::{
 };
 use crate::base::rdata::RtypeRecordData;
 #[cfg(feature = "master")]
-use crate::master::scan::{CharSource, Scan, ScanError, Scanner};
+use crate::master::scan::{CharSource, Scanner};
+#[cfg(feature = "scan")]
+use crate::scan::{Scan, ScanError};
 use crate::utils::base64;
 #[cfg(feature = "master")]
 use bytes::Bytes;
